Renamed config to develop from dev
